requirements.txt
pdfplumber==0.10.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
mistralai==0.4.0
numpy==1.24.3
